+++ assets-src/rhau-admin.js
@@ -7,26 +7,36 @@ export default class RHAU {
 
   constructor() {
     jQuery(document).ready(() => this.onDocReady());
+    this.reopenAcfFieldObjects();
     this.restoreScrollTop();
+    $('form#post').one( 'submit', (e) => this.beforeSubmitPostForm(e) );
+    
   }
 
   /**
-   * Is run on doc ready
+   * This runs on doc ready
    */
   onDocReady() {
-    this.restoreScrollTop(true);
     this.enableSubmitDiv();
-    $('form#post').submit((e) => this.storeScrollTop());
     this.initAdminBarButtons();
+    this.restoreScrollTop();
+    this.removeFromStore('scrollTop');
+  }
+
+  /**
+   * Is being fired before saving/publishing/updating a post
+   */
+  beforeSubmitPostForm(e) {
+    // e.preventDefault();
+    this.storeOpenAcfFieldObjects();
+    this.addToStore('scrollTop', $(document).scrollTop());
   }
 
   /**
    * Restores scrollTop
-   * @param {boolean} deleteValue 
    */
-  restoreScrollTop(deleteValue = false) {
-    var scrollTop = sessionStorage.getItem(this.getStorageKey('scrollTop'));
-    if( deleteValue ) sessionStorage.removeItem(this.getStorageKey('scrollTop'));
+  restoreScrollTop() {
+    var scrollTop = this.getFromStore('scrollTop');
     if( scrollTop ) $(document).scrollTop(parseInt(scrollTop));
   }
 
@@ -64,17 +74,79 @@ export default class RHAU {
   }
 
   /**
-   * Store current scrollTop
+   * Stores something in session storage
+   * 
+   * @param {string} key 
+   * @param {mixed} value 
    */
-  storeScrollTop() {
-    sessionStorage.setItem(this.getStorageKey('scrollTop'), $(document).scrollTop());
+  addToStore(key, value) {
+    sessionStorage.setItem(this.getStorageKey(key), JSON.stringify(value));
+  }
+
+  /**
+   * Removes something from session storage
+   * @param {string} key 
+   */
+  removeFromStore(key) {
+    sessionStorage.removeItem(this.getStorageKey(key));
+  }
+
+  /**
+   * Gets something from store
+   * @param {string} key 
+   */
+  getFromStore(key) {
+    let value = sessionStorage.getItem(this.getStorageKey(key));
+    return value ? JSON.parse(value) : value;
+  }
+
+  /**
+   * Stores open Field objects
+   */
+  storeOpenAcfFieldObjects() {
+    if( typeof acf.getFieldObjects !== 'function' ) return;
+    let openFieldObjects = [];
+    try {
+      for( const fieldObject of acf.getFieldObjects() ) {
+        if( !fieldObject.isOpen() ) continue;
+        // disable close function
+        fieldObject.close = () => {};
+        openFieldObjects.push( fieldObject.getKey() );
+      }
+    } catch(e) { console.warn(e) }
+    this.addToStore('open-acf-field-objects', openFieldObjects);
+  }
+
+  /**
+   * Restores open field objects
+   */
+  reopenAcfFieldObjects() {
+    if( typeof acf.getFieldObjects !== 'function' ) return;
+    let openObjects = this.getFromStore('open-acf-field-objects');
+    if( !openObjects ) return;
+    try {
+      for( const key of openObjects ) {
+        const fieldObject = acf.getFieldObject(key);
+        const $settings = fieldObject.$el.children('.settings');
+        // copied code from acf-field-group.js
+        // open
+        $settings.slideDown(0);
+        fieldObject.$el.addClass('open');
+        acf.doAction('open_field_object', fieldObject);
+        fieldObject.trigger('openFieldObject');
+        // action (show)
+        acf.doAction('show', $settings);
+      } 
+    } catch(e) { console.warn(e) }
+    // removes the property from store
+    this.removeFromStore('open-acf-field-objects');
   }
 
   /**
    * Get storage key for scrollTop
    */
   getStorageKey(key) {
-    var path = window.location.pathname + window.location.search;
+    var path = window.location.pathname;
     return key + ":" + path.replace(/^\/+/g, '').split('/').join('-');
   }